+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * dd86a11 - 2017-10-09: Fixed issue with API help command 
+- * f790c1f - 2017-10-09: Fixed issues with API multi commands 
+- * e4be7cf - 2017-10-09: Fixed issue with API help command 
 - * 68b8996 - 2017-10-08: Fixed API builder 
 - * a455d99 - 2017-10-08: Fixed API cmds 
 - * 12b6e6b - 2017-10-07: Release v9.1.31